Some trigger setup and directory rearrangement
