+++ Common/Commands/ExampleChatCommand.cs
@@ -15,7 +15,8 @@ public class ChatTriggerExample : ModCommand
     // A short description of this command
     public override string Description => "developer testing command";
 
-    public override void Action(CommandCaller caller, string input, string[] args) {
+    public override void Action(CommandCaller caller, string input, string[] args)
+    {
         Main.NewText("Triggered Chat Command");
         Main.NewText("Display Width: " + Main.screenWidth.ToString());
         Main.NewText("Display Height: " + Main.screenHeight.ToString());
