+++ FresherFishing.csproj
@@ -1,14 +1,14 @@
 <?xml version="1.0" encoding="utf-8"?>
 <Project Sdk="Microsoft.NET.Sdk">
-  <Import Project="..\tModLoader.targets" />
-  <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
-    <PlatformTarget>AnyCPU</PlatformTarget>
-    <LangVersion>latest</LangVersion>
-    <PackageId>Fresher-Fishing</PackageId>
-    <RootNamespace>Fresher-Fishing</RootNamespace>
-  </PropertyGroup>
-  <ItemGroup>
-    <PackageReference Include="tModLoader.CodeAssist" Version="0.1.*" />
-  </ItemGroup>
+    <Import Project="..\tModLoader.targets" />
+    <PropertyGroup>
+        <TargetFramework>net6.0</TargetFramework>
+        <PlatformTarget>AnyCPU</PlatformTarget>
+        <LangVersion>latest</LangVersion>
+        <PackageId>Fresher-Fishing</PackageId>
+        <RootNamespace>Fresher-Fishing</RootNamespace>
+    </PropertyGroup>
+    <ItemGroup>
+        <PackageReference Include="tModLoader.CodeAssist" Version="0.1.*" />
+    </ItemGroup>
 </Project>
